+++ packages/editor/src/RichContentEditor/RichContentEditor.jsx
@@ -15,6 +15,7 @@ import handleKeyCommand from './handleKeyCommand';
 import handleReturnCommand from './handleReturnCommand';
 import blockStyleFn from './blockStyleFn';
 import getBlockRenderMap from './getBlockRenderMap';
+import { combineStyleFns } from './combineStyleFns';
 import { getStaticTextToolbarId } from './Toolbars/toolbar-id';
 import {
   AccessibilityListener,
@@ -67,10 +68,25 @@ class RichContentEditor extends Component {
   };
 
   initPlugins() {
-    const { helpers, plugins, config, isMobile, anchorTarget, relValue, t } = this.props;
+    const {
+      helpers,
+      plugins,
+      config,
+      isMobile,
+      anchorTarget,
+      relValue,
+      t,
+      customStyleFn,
+    } = this.props;
 
     const { theme } = this.state;
-    const { pluginInstances, pluginButtons, pluginTextButtons, pubsubs } = createPlugins({
+    const {
+      pluginInstances,
+      pluginButtons,
+      pluginTextButtons,
+      pubsubs,
+      pluginStyleFns,
+    } = createPlugins({
       plugins,
       config,
       helpers,
@@ -86,6 +102,7 @@ class RichContentEditor extends Component {
     this.pluginKeyBindings = initPluginKeyBindings(pluginTextButtons);
     this.plugins = [...pluginInstances, ...Object.values(this.toolbars)];
     this.subscriberPubsubs = pubsubs || [];
+    this.customStyleFn = combineStyleFns([...pluginStyleFns, customStyleFn]);
   }
 
   initEditorToolbars(pluginButtons, pluginTextButtons) {
@@ -117,6 +134,7 @@ class RichContentEditor extends Component {
       refId: this.refId,
       getToolbarSettings: config.getToolbarSettings,
       uiSettings: config.uiSettings,
+      config,
     });
   }
 
@@ -294,6 +312,7 @@ class RichContentEditor extends Component {
         )}
         editorKey={editorKey}
         keyBindingFn={keyBindingFn(this.getCustomCommandHandlers().commands || [])}
+        customStyleFn={this.customStyleFn}
         helpers={helpers}
         tabIndex={tabIndex}
         placeholder={placeholder || ''}
@@ -384,12 +403,14 @@ RichContentEditor.propTypes = {
   handleBeforeInput: PropTypes.func,
   handlePastedText: PropTypes.func,
   handleReturn: PropTypes.func,
+  customStyleFn: PropTypes.func,
   locale: PropTypes.string,
 };
 
 RichContentEditor.defaultProps = {
   config: {},
   spellCheck: true,
+  customStyleFn: () => ({}),
 };
 
 export default translate(null, { withRef: true })(RichContentEditor);